+++ extensions/open-path/src/open-path.tsx
@@ -1,9 +1,10 @@
 import { getPreferenceValues, open, showHUD, showInFinder } from "@raycast/api";
 import { fetchItemInputClipboardFirst, fetchItemInputSelectedFirst } from "./utils/input-item";
-import { checkIsFile, isEmpty, isUrl, Preference, searchUrlBuilder, urlBuilder } from "./utils/common-utils";
+import { checkIsFile, isEmail, isEmpty, isUrl, Preference, searchUrlBuilder, urlBuilder } from "./utils/common-utils";
 import fse from "fs-extra";
 import { homedir } from "os";
 import { URL } from "url";
+import { isIPv4 } from "net";
 
 export default async () => {
   const { trimText, isShowHud, priorityDetection, searchEngine } = getPreferenceValues<Preference>();
@@ -15,16 +16,34 @@ export default async () => {
     await showHud(isShowHud, "No path detected");
     return "";
   }
-
   if (fse.pathExistsSync(path)) {
     await open(path);
-    console.info("open(path)");
+    console.info("open: directory path");
+    await showHud(isShowHud, "Open Path: " + path);
+    return;
+  }
+
+  if (isIPv4(path)) {
+    if (path == "127.0.0.1") {
+      await open("http://www." + path);
+    } else {
+      await open("http://" + path);
+    }
+    console.info("open: IP address");
+    await showHud(isShowHud, "Open IP: " + path);
+    return;
+  }
+
+  if (isEmail(path)) {
+    await open("mailto:" + path);
+    console.info("open: email");
+    await showHud(isShowHud, "Send Email: " + path);
     return;
   }
 
   try {
     await open(new URL(path).toString());
-    console.info("open(new URL(path).toString())");
+    console.info("open: URL");
     await showHud(isShowHud, "Open URL: " + path);
   } catch (e) {
     await reOpenURL(isShowHud, path, searchEngine);