+++ extensions/open-path/src/utils/common-utils.ts
@@ -22,6 +22,11 @@ export const checkIsFile = (path: string) => {
   }
 };
 
+export function isEmail(text: string): boolean {
+  const regex = /^[\da-zA-Z_.-]+@[\da-zA-Z_.-]+([.][a-zA-Z]+){1,2}$/;
+  return regex.test(text);
+}
+
 export function isUrl(text: string): boolean {
   const regex = /^((http|https|ftp):\/\/)?((?:[\w-]+\.)+[a-z\d]+)((?:\/[^/?#]*)+)?(\?[^#]+)?(#.+)?$/i;
   return regex.test(text) || isIP(text);